Support Doxygen automated code documentation
Doxygen is the de factor standard tool for generating documentation
from annotated C++ sources, but it also supports other popular
programming languages.

Run with 'make doxygen', output in docs/doxygen/

The outputs are not tracked in Git, as the intent is that developers
would locally generate documentation from the source code as is.
To track in Git ignores the automated nature of the documentation
generation that Doxygen provides.

In the future a centrally hosted Doxygen output may be provided.