Improve cross platform support for data/preload.lua
I was looking for the forward and backward slashes in hardcoded data file
paths the engine looks for, and came across this one.

Whilst reviewing the code, there's a better fix as this function already
chdir()'d into $SOULRIDE_TOP/data so we don't need to walk up to the
parent directory and then back into $SOULRIDE_TOP/data to find that file.

This completely obviates the forward/backward slash issue here.

Tested with game data that the global $SOULRIDE_TOP/data/preload.lua is
found and ran within the engine before and after this change.